[CHORE/#55] 필요없는 Log 삭제, JwtResponse kakaoId 변수 타입 변경

+++ app/src/main/java/umc/link/zip/data/dto/response/JwtResponse.kt
@@ -5,7 +5,7 @@ import umc.link.zip.domain.model.login.JwtModel
 data class JwtResponse(
     val userId: Int,
     val nickname: String,
-    val kakaoId: Int,
+    val kakaoId: Long,
     val createdAt: String
 ) {
     fun toModel() = JwtModel(
